Do not delete torchvision

+++ setupbase.py
@@ -285,7 +285,7 @@ class BuildExtRelocate(BuildExtension):
         }
         relocate, library_path = relocation_funcs[sys.platform]
 
-        if osp.exists(library_path):
+        if osp.exists(library_path) and library_path != base_library_dir:
             shutil.rmtree(library_path)
             os.makedirs(library_path)
